--- collector/env_samples/devnet.env
@@ -1,3 +1,0 @@
-WALLETS_CONFIG="./collector/testdata/wallets#foo.config.json"
-AUTH_REGISTRATION="./collector/testdata/auth#foo.devnet.json"
-RECEIVER="erd1testnlersh4z0wsv8kjx39me4rmnvjkwu8dsaea7ukdvvc9z396qykv7z7"
