+++ wizard/collect_rewards.py
@@ -6,13 +6,13 @@ from pathlib import Path
 
 from rich import print
 
-from collector import errors, ux
-from collector.accounts import load_accounts
-from collector.configuration import CONFIGURATIONS
-from collector.entrypoint import MyEntrypoint
-from collector.errors import UsageError
-from collector.rewards import ReceivedRewardsOfAccount
-from collector.utils import format_time
+from wizard import errors, ux
+from wizard.accounts import load_accounts
+from wizard.configuration import CONFIGURATIONS
+from wizard.entrypoint import MyEntrypoint
+from wizard.errors import UsageError
+from wizard.rewards import ReceivedRewardsOfAccount
+from wizard.utils import format_time
 
 
 def main(cli_args: list[str] = sys.argv[1:]):
