+++ wizard/env_samples/devnet.env
@@ -1,0 +1,3 @@
+WALLETS_CONFIG="./wizard/testdata/wallets#foo.config.json"
+AUTH_REGISTRATION="./wizard/testdata/auth#foo.devnet.json"
+RECEIVER="erd1testnlersh4z0wsv8kjx39me4rmnvjkwu8dsaea7ukdvvc9z396qykv7z7"
