+++ wizard/guardians_status.py
@@ -5,11 +5,11 @@ from pathlib import Path
 
 from rich import print
 
-from collector import errors, ux
-from collector.accounts import load_accounts
-from collector.configuration import CONFIGURATIONS
-from collector.entrypoint import MyEntrypoint
-from collector.guardians import AuthApp
+from wizard import errors, ux
+from wizard.accounts import load_accounts
+from wizard.configuration import CONFIGURATIONS
+from wizard.entrypoint import MyEntrypoint
+from wizard.guardians import AuthApp
 
 
 def main(cli_args: list[str] = sys.argv[1:]):
