+++ wizard/guardians_update.py
@@ -3,16 +3,16 @@ import traceback
 from argparse import ArgumentParser
 from pathlib import Path
 
+from multiversx_sdk import Address
 from rich import print
-from rich.prompt import Confirm
 from rich.rule import Rule
 
-from collector import errors, ux
-from collector.accounts import AccountWrapper, load_accounts
-from collector.configuration import CONFIGURATIONS
-from collector.entrypoint import MyEntrypoint
-from collector.guardians import AuthApp
-from collector.transactions import TransactionWrapper
+from wizard import errors, ux
+from wizard.accounts import AccountWrapper, load_accounts
+from wizard.configuration import CONFIGURATIONS
+from wizard.entrypoint import MyEntrypoint
+from wizard.guardians import AuthApp
+from wizard.transactions import TransactionWrapper
 
 
 def main(cli_args: list[str] = sys.argv[1:]):
@@ -28,25 +28,27 @@ def _do_main(cli_args: list[str]):
     parser = ArgumentParser()
     parser.add_argument("--network", choices=CONFIGURATIONS.keys(), required=True, help="network name")
     parser.add_argument("--wallets", required=True, help="path of the wallets configuration file")
-    parser.add_argument("--auth", required=True, help="auth registration file")
+    parser.add_argument("--new-auth", required=True, help="auth registration file")
     args = parser.parse_args(cli_args)
 
     network = args.network
     configuration = CONFIGURATIONS[network]
     entrypoint = MyEntrypoint(configuration)
     accounts_wrappers = load_accounts(Path(args.wallets))
-    auth_app = AuthApp.new_from_registration_file(Path(args.auth))
+    new_auth_app = AuthApp.new_from_registration_file(Path(args.new_auth))
+    empty_auth_app = AuthApp([])
 
     accounts_wrappers_by_addresses: dict[str, AccountWrapper] = {
         item.account.address.to_bech32(): item for item in accounts_wrappers
     }
 
     entrypoint.recall_nonces(accounts_wrappers)
+    entrypoint.recall_guardians(accounts_wrappers)
     transactions_wrappers: list[TransactionWrapper] = []
 
-    ux.show_message("Creating and signing 'guard account' transactions for all auth registration entries...")
+    ux.show_message("Creating and signing 'set (update) guardian' transactions for all auth registration entries...")
 
-    for entry in auth_app.get_all_entries():
+    for entry in new_auth_app.get_all_entries():
         account_wrapper = accounts_wrappers_by_addresses.get(entry.get_address())
         if not account_wrapper:
             raise errors.UsageError(f"account (wallet) not found for registration entry {entry.get_address()}")
@@ -58,28 +60,18 @@ def _do_main(cli_args: list[str]):
         print(Rule())
         print(address.to_bech32(), f"([yellow]{label}[/yellow])")
 
-        guardian_data = entrypoint.get_guardian_data(address)
+        existing_guardian_data = entrypoint.get_guardian_data(address)
 
-        if guardian_data.is_guarded:
-            print(f"... account is [blue]already guarded[/blue]")
-            if not Confirm.ask("Re-guard it?"):
-                continue
+        if not existing_guardian_data.is_guarded:
+            print(f"... account is [blue]not guarded[/blue], will be skipped")
+            continue
 
-        if not guardian_data.active_guardian:
-            print(f"... account has [red]no active guardian[/red]")
-            if not Confirm.ask("Attempt to guard (won't work)?"):
-                continue
-
-        if entry.get_guardian() != guardian_data.active_guardian:
-            print(f"... registered guardian [red]does not match[/red] the active guardian")
-            if not Confirm.ask("Attempt to guard (please don't)?"):
-                continue
-
-        transaction = entrypoint.guard_account(account_wrapper)
+        new_guardian = Address.new_from_bech32(entry.get_guardian())
+        transaction = entrypoint.set_guardian(account_wrapper, new_guardian)
         transactions_wrappers.append(TransactionWrapper(transaction, label))
 
-    ux.confirm_continuation(f"Ready to guard accounts, by sending [green]{len(transactions_wrappers)}[/green] transactions?")
-    entrypoint.send_multiple(auth_app, transactions_wrappers)
+    ux.confirm_continuation(f"Ready to update guardians, by sending [green]{len(transactions_wrappers)}[/green] transactions?")
+    entrypoint.send_multiple(empty_auth_app, transactions_wrappers)
 
 
 if __name__ == "__main__":
