+++ wizard/testdata/auth#bar.testnet.json
