+++ wizard/testdata/auth#empty.json
@@ -1,0 +1,1 @@
+[]
