+++ wizard/testdata/auth#foo.devnet.json
