+++ wizard/testdata/auth#foo.testnet.json
