+++ wizard/testdata/wallets#bar.config.json
