+++ wizard/testdata/wallets#foo.config.json
