+++ wizard/testdata/wallets.config.json
